Moving views/pages into folders related to usage.
