+++ Examples/MAUI-gRPC/MauiAntClientApp/AppShell.xaml.cs
@@ -1,5 +1,7 @@
 ﻿using MauiAntClientApp.Views;
-using MauiAntClientApp.Views.FitnessEquipment;
+using MauiAntClientApp.Views.FitnessEquipmentPages;
+using MauiAntClientApp.Views.GeocachePages;
+using MauiAntClientApp.Views.HeartRatePages;
 
 namespace MauiAntClientApp
 {
