+++ Examples/MAUI-gRPC/MauiAntClientApp/Views/FitnessEquipmentPages/NordicSkierView.xaml
@@ -2,14 +2,13 @@
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:ant="clr-namespace:SmallEarthTech.AntPlus.DeviceProfiles.FitnessEquipment;assembly=SmallEarthTech.AntPlus"
-             x:DataType="ant:Elliptical"
-             x:Class="MauiAntClientApp.Views.FitnessEquipment.EllipticalView">
+             x:DataType="ant:NordicSkier"
+             x:Class="MauiAntClientApp.Views.FitnessEquipmentPages.NordicSkierView">
     <VerticalStackLayout>
-        <Label Text="Elliptical" FontAttributes="Bold" TextDecorations="Underline"/>
+        <Label Text="Nordic Skier" FontAttributes="Bold" TextDecorations="Underline"/>
         <Label Text="{Binding Cadence, StringFormat='Cadence: {0}'}"/>
         <Label Text="{Binding InstantaneousPower, StringFormat='Instantaneous Power: {0}'}"/>
         <Label Text="{Binding StrideCount, StringFormat='Stride Count: {0}'}"/>
-        <Label Text="{Binding PosVerticalDistance, StringFormat='PosVerticalDistance: {0}'}"/>
         <Label Text="{Binding Capabilities, StringFormat='Capabilities: {0}'}"/>
     </VerticalStackLayout>
 </ContentView>
