+++ Examples/MAUI-gRPC/MauiAntClientApp/Views/FitnessEquipmentPages/NordicSkierView.xaml.cs
@@ -1,6 +1,6 @@
 using SmallEarthTech.AntPlus.DeviceProfiles.FitnessEquipment;
 
-namespace MauiAntClientApp.Views.FitnessEquipment;
+namespace MauiAntClientApp.Views.FitnessEquipmentPages;
 
 public partial class NordicSkierView : ContentView
 {
