+++ Examples/MAUI-gRPC/MauiAntClientApp/Views/FitnessEquipmentPages/RowerView.xaml
@@ -2,13 +2,13 @@
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:ant="clr-namespace:SmallEarthTech.AntPlus.DeviceProfiles.FitnessEquipment;assembly=SmallEarthTech.AntPlus"
-             x:DataType="ant:NordicSkier"
-             x:Class="MauiAntClientApp.Views.FitnessEquipment.NordicSkierView">
+             x:DataType="ant:Rower"
+             x:Class="MauiAntClientApp.Views.FitnessEquipmentPages.RowerView">
     <VerticalStackLayout>
-        <Label Text="Nordic Skier" FontAttributes="Bold" TextDecorations="Underline"/>
+        <Label Text="Rower" FontAttributes="Bold" TextDecorations="Underline"/>
         <Label Text="{Binding Cadence, StringFormat='Cadence: {0}'}"/>
         <Label Text="{Binding InstantaneousPower, StringFormat='Instantaneous Power: {0}'}"/>
-        <Label Text="{Binding StrideCount, StringFormat='Stride Count: {0}'}"/>
+        <Label Text="{Binding StrokeCount, StringFormat='Stroke Count: {0}'}"/>
         <Label Text="{Binding Capabilities, StringFormat='Capabilities: {0}'}"/>
     </VerticalStackLayout>
 </ContentView>
