+++ Examples/MAUI-gRPC/MauiAntClientApp/Views/FitnessEquipmentPages/TreadmillView.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:ant="clr-namespace:SmallEarthTech.AntPlus.DeviceProfiles.FitnessEquipment;assembly=SmallEarthTech.AntPlus"
              x:DataType="ant:Treadmill"
-             x:Class="MauiAntClientApp.Views.FitnessEquipment.TreadmillView">
+             x:Class="MauiAntClientApp.Views.FitnessEquipmentPages.TreadmillView">
     <VerticalStackLayout>
         <Label Text="Treadmill" FontAttributes="Bold" TextDecorations="Underline"/>
         <Label Text="{Binding Cadence, StringFormat='Cadence: {0}'}"/>
