+++ Examples/MAUI-gRPC/MauiAntClientApp/Views/GeocachePages/ProgramGeocacheView.xaml
@@ -2,7 +2,7 @@
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              x:Name="this"
-             x:Class="MauiAntClientApp.Controls.ProgramGeocacheView">
+             x:Class="MauiAntClientApp.Views.GeocachePages.ProgramGeocacheView">
     <Frame>
         <VerticalStackLayout BindingContext="{x:Reference this}">
             <Grid RowDefinitions="*, *, *, *" ColumnDefinitions="Auto, *">
