+++ Examples/MAUI-gRPC/MauiAntClientApp/Views/HeartRatePages/HeartRatePage.xaml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:MauiAntClientApp.Controls"
              xmlns:viewmodels="clr-namespace:MauiAntClientApp.ViewModels"
+             xmlns:views="clr-namespace:MauiAntClientApp.Views.HeartRatePages"
              x:DataType="viewmodels:HeartRateViewModel"
-             x:Class="MauiAntClientApp.Views.HeartRatePage"
+             x:Class="MauiAntClientApp.Views.HeartRatePages.HeartRatePage"
              Title="HeartRatePage">
     <Grid Margin="10" RowDefinitions="Auto, *">
         <HorizontalStackLayout Spacing="24">
@@ -49,7 +49,7 @@
                         </StackLayout>
                     </Frame>
                 </VerticalStackLayout>
-                <controls:HeartRateCapabilitiesView HeartRateViewModel="{Binding .}" IsVisible="{Binding Source={x:Reference cap}, Path=IsChecked}"/>
+                <views:HeartRateCapabilitiesView HeartRateViewModel="{Binding .}" IsVisible="{Binding Source={x:Reference cap}, Path=IsChecked}"/>
             </VerticalStackLayout>
         </ScrollView>
